rag_course_notes_chatbot/requirements.txt
# Python dependencies for RAG Course Notes Chatbot
langchain>=0.3.0,<0.4.0
gradio>=5.0.0,<6.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
pillow>=10.0.0
faiss-cpu>=1.7.4
matplotlib>=3.8.0
pandas>=2.1.0
numpy>=1.24.0
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
jupyter>=1.0.0
ipykernel>=6.27.0
tqdm>=4.66.0
scikit-learn>=1.3.0
seaborn>=0.13.0